fix(api): stop parsing a double as a long

+++ commons/src/main/java/com/envyful/api/type/UtilParse.java
@@ -101,6 +101,6 @@ public class UtilParse {
             return OptionalDouble.empty();
         }
 
-        return OptionalDouble.of(Long.parseLong(arg));
+        return OptionalDouble.of(Double.parseDouble(arg));
     }
 }
